requirements.txt
flask
flask-cors
supabase
python-dotenv
requests
APScheduler